Refactor command classes into package
