--- src/main/java/seedu/address/logic/commands/AddMasteryCheckCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.commands;
-
-public class AddMasteryCheckCommand {
-}
