+++ src/main/java/seedu/address/logic/commands/add/AddMasteryCheckCommand.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands.add;
+
+public class AddMasteryCheckCommand {
+}
