+++ src/main/java/seedu/address/logic/commands/view/ViewUpcomingConsultationsCommand.java
@@ -1,8 +1,9 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.view;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_UPCOMING_CONSULTATIONS;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 
 /**
@@ -14,7 +15,7 @@ public class ViewUpcomingConsultationsCommand extends ViewCommand {
     @Override
     public CommandResult execute(Model model) {
         requireNonNull(model);
-        model.updateFilteredConsultationsList(PREDICATE_SHOW_UPCOMING_CONSULTATIONS);
+        model.updateConsultationsList(PREDICATE_SHOW_UPCOMING_CONSULTATIONS);
 
         return new CommandResult(MESSAGE_SUCCESS);
     }
